backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.7.2
python-dotenv==1.0.1
